+++ onsiteprint-plugin/blocks/event/block-script-parts/setup-header.js
@@ -1,8 +1,8 @@
 /* ------------------------------------------------------------------------
  #  JS Part Name: Setup Header
  *  Block function included in the Event Block.
- ?  Updated: 2025-08-04 - 04:31 (Y:m:d - H:i)
- ?  Info: Changed how the Modal Toggle Listener is used.
+ ?  Updated: 2025-08-06 - 04:00 (Y:m:d - H:i)
+ ?  Info: Added new Modal Toggle Listener to new Cancel Error Button.
 
 ---------------------------------------------------------------------------
  #  TABLE OF CONTENTS:
@@ -16,6 +16,7 @@
  #  1. Import Functions from Scripts
 --------------------------------------------------------------------------- */
 import * as opModuleBasic from '../../../assets/js/inc/basic.js'
+import { opModalToggleListener } from '../../../assets/js/inc/modal/toggle-modal-listener.js'
 import { opModalCreateParticipant } from './modals/modal-create-participant.js'
 import * as opModuleParticipantListeners from './participant-listeners.js'
 
@@ -37,12 +38,16 @@ export function opSetupHeader( debug, block, eventId, columnAmount ) {
 
         ///// Add Create Participant Modal.
         opModalCreateParticipant( debug, block, eventId )
-            
+
         ///// Get the Button Elements.
+        let cancelButton = block.querySelector( '.op-cancel_error' )
         let csvButton = block.querySelector( '.op-button-csv' )
 
+        ///// Set Modal Toggle Listener to the Cancel Error Button.
+        opModalToggleListener( debug, cancelButton, false )
+
         ///// Set Modal Toggle Listener to the CSV Download Button.
-        //opModalToggleListener( debug, csvButton, true, opModuleParticipantListeners.opDownloadCSVFileListener( debug, block, csvButton, eventId ) )
+        opModalToggleListener( debug, csvButton, true, opModuleParticipantListeners.opDownloadCSVFileListener( debug, block, csvButton, eventId ) )
 
         ///// Return the Response.
         return opModuleBasic.opReturnResponse( false, 200, {