+++ onsiteprint-plugin/blocks/event/block-template-parts/modal.php
@@ -1,8 +1,8 @@
 <?php
 /* ------------------------------------------------------------------------
  *  Block Part Name: Modal
- ?  Updated: 2025-07-27 - 04:52 (Y:m:d - H:i)
- ?  Info: Added new Error Button.
+ ?  Updated: 2025-08-06 - 04:06 (Y:m:d - H:i)
+ ?  Info: Added new Cancel Error Button and moved Close Button.
 ---------------------------------------------------------------------------
  #  The Block Part Content
 --------------------------------------------------------------------------- */
@@ -15,6 +15,11 @@
 
             <div class="op-modal-error">
                 <div class="op-modal-error__inner">
+
+                    <button type="button" class="op-button-close op-cancel_error op-button op-button-size-small op-button-style-outline" data-color="primary-90" data-icon="xmark" data-icon-position="right" data-title-visibility="1">
+                        <span class="op-icon" role="img" aria-label="X Mark Icon"></span>
+                        <span class="op-button-title"><?= esc_attr( $modal['cancel_button'] ) ?></span>
+                    </button>
 
                     <h3 class="op-modal-title" data-icon="circle-exclamation">
                         <span class="op-icon" role="img" aria-label="Error Icon"></span>
@@ -31,11 +36,6 @@
             </div>
 
             <div class="op-header-content"></div>
-
-            <button type="button" class="op-button-cancel op-button op-button-size-small op-button-style-outline" data-color="primary-90" data-icon="xmark" data-icon-position="right" data-title-visibility="1">
-                <span class="op-icon" role="img" aria-label="X Mark Icon"></span>
-                <span class="op-button-title"><?= esc_attr( $modal['cancel_button'] ) ?></span>
-            </button>
 
         </div>